added a warehouselist component and added routes to warehosues component
