+++ package.json
@@ -9,11 +9,9 @@
     "axios": "^0.27.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-
     "react-router-dom": "^5.3.3",
     "react-scripts": "5.0.1",
     "sass": "^1.53.0",
-
     "web-vitals": "^2.1.4"
   },
   "scripts": {
