+++ src/App.js
@@ -4,10 +4,13 @@ import { BrowserRouter } from "react-router-dom";
 import { Route } from "react-router-dom";
 import { Switch } from "react-router-dom";
 // import Header from "./components/Header/Header";
-import Warehouse from "./pages/Warehouses/Warehouses";
-// import Inventory from "./pages/Inventory/Inventory";
+// import warehouses from "./pages/Warehouses/Warehouses";
+import Inventory from "./pages/Inventory/Inventory";
 // import Footer from "./components/Footer/Footer";
-import WarehouseDetails from "./components/WarehouseDetails/WarehouseDetails";
+// import WarehouseDetails from "./components/WarehouseDetails/WarehouseDetails";
+// import Warehouses from "./components/WarehouseList/WarehouseList";
+import Warehouses from "./pages/Warehouses/Warehouses";
+import Footer from "./components/Footer/Footer"
 
 class App extends React.Component {
   render() {
@@ -16,20 +19,18 @@ class App extends React.Component {
         <BrowserRouter>
           {/* <Header /> */}
           <Switch>
-            <Route path="/" exact component={Warehouse} />
-
-            <Route path="/warehouse" exact component={Warehouse} />
+            <Route path="/" exact component={Warehouses} />
+            <Route path="/inventory" component={Inventory} />
+            {/* <Route path="/warehouse" exact component={Warehouses} />
             <Route path="/warehouse/:id" exact component={WarehouseDetails} />
             <Route path="/warehouse/add-new-warehouse" component={Warehouse} />
             <Route path="/warehouse/edit-warehouse" component={Warehouse} />
-            <Route path="/inventory" component={Inventory} />
-            <Route path="/inventory/:id" component={Inventory} />
-
+            <Route path="/inventory/:id" component={Inventory} /> */}
             <Route
               component={() => alert("(404) Bad Request, Page Not Found")}
             />
           </Switch>
-          {/* <Footer /> */}
+          <Footer />
         </BrowserRouter>
       </div>
     );
